chore: added systemMessage type
