--- packages/types/src/card.ts
@@ -1,1356 +1,0 @@
-/* tslint:disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-import { ServerAction, DeepLinkAction, TextAction } from './response';
-
-export type Card = DiscoveryCard | GalleryCard | GridCard | ListCard | TitleCard;
-/**
- * Карточка с вертикальным списком ячеек для отображения в ленте.
- */
-export type DiscoveryCard = CardBase & {
-    /**
-     * Ячейки с контентом карточки
-     */
-    cells: {
-        [k: string]: unknown;
-    };
-    background_image?: ImageView;
-    /**
-     * Цвет, которым заливается фон карточки. Задаётся как HEX c компонентом alpha в виде #AARRGGBB или #RRGGBB
-     */
-    background_color?: string;
-    /**
-     * Имеет ли карточка затемнение в нижней части
-     */
-    has_fade?: boolean;
-    /**
-     * Количество колонок, которое занимает карточка по ширине
-     */
-    width_columns?: 2 | 3 | 4;
-    /**
-     * Соотношение сторон карточки - W:H
-     */
-    aspect_ratio?: '1:1' | '2:3' | '3:2' | '3:4' | '16:9';
-    actions?: Actions;
-    type: 'discovery_card';
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-};
-/**
- * Адрес картинки в локальном пространстве клиента https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=19%3A9
- */
-export type LocalImageIdentificator =
-    | 'add_contactless_payment'
-    | 'apple_wallet'
-    | 'arrow_right_chevron_rectangle'
-    | 'arrow_right_trace_line'
-    | 'arrows_forward_back'
-    | 'atm_banknote'
-    | 'bag'
-    | 'ballot_paper'
-    | 'bar_graph'
-    | 'basket'
-    | 'beach_umbrella'
-    | 'bell'
-    | 'bell_crossed'
-    | 'between_accounts'
-    | 'boarding_pass'
-    | 'bolt'
-    | 'box'
-    | 'building'
-    | 'calculator'
-    | 'camera'
-    | 'candy'
-    | 'capsule'
-    | 'car'
-    | 'card'
-    | 'card_amex'
-    | 'card_amex_black'
-    | 'card_amex_gold'
-    | 'card_blocked'
-    | 'card_chevron_left'
-    | 'card_ghost'
-    | 'card_maestro_null'
-    | 'card_mc'
-    | 'card_mc_black'
-    | 'card_mc_gold'
-    | 'card_mc_null'
-    | 'card_mir'
-    | 'card_mir_black'
-    | 'card_mir_gold'
-    | 'card_mir_null'
-    | 'card_momentum'
-    | 'card_on_card'
-    | 'card_plus'
-    | 'card_pro100_null'
-    | 'card_uek'
-    | 'card_viewfinder'
-    | 'card_visa_gold'
-    | 'card_visa_black'
-    | 'card_visa'
-    | 'card_visa_infinite_sber'
-    | 'card_visa_infinite'
-    | 'card_visa_digital'
-    | 'card_visa_null'
-    | 'cards_carousel'
-    | 'case'
-    | 'case_diagram'
-    | 'chess_piece'
-    | 'circle_cross'
-    | 'circle_plus'
-    | 'clock'
-    | 'coat_of_arms'
-    | 'coin'
-    | 'coin_heart'
-    | 'color_spasibo'
-    | 'color_total'
-    | 'counterclock_wise'
-    | 'crown'
-    | 'digital_banknotes'
-    | 'document'
-    | 'document_checkmark'
-    | 'document_magnifying_glass'
-    | 'document_on_document'
-    | 'drop'
-    | 'ellipsis'
-    | 'eye'
-    | 'eye_crossed'
-    | 'face'
-    | 'face_id'
-    | 'falling_coin'
-    | 'file_pdf'
-    | 'fingerprint'
-    | 'flag'
-    | 'flame'
-    | 'flasher'
-    | 'funnel'
-    | 'gamepad'
-    | 'gear'
-    | 'gift'
-    | 'globe'
-    | 'hammer'
-    | 'handset'
-    | 'heart'
-    | 'heart_hand'
-    | 'house'
-    | 'house_handset'
-    | 'house_percent'
-    | 'info'
-    | 'key'
-    | 'list'
-    | 'lock_closed'
-    | 'lock_opened'
-    | 'magnit'
-    | 'mail'
-    | 'mail_eye'
-    | 'mail_waves'
-    | 'man_badge'
-    | 'man_wheelchair'
-    | 'map'
-    | 'mask'
-    | 'mindmap'
-    | 'mobile'
-    | 'mobile_plus'
-    | 'mobile_plus_alt'
-    | 'mobile_waves'
-    | 'necklace'
-    | 'note'
-    | 'notebook_waves'
-    | 'open_book'
-    | 'other'
-    | 'paper_tray'
-    | 'parking'
-    | 'passport'
-    | 'paw'
-    | 'pencil_line'
-    | 'percent'
-    | 'pfm'
-    | 'photo'
-    | 'pie_chart'
-    | 'placeholder'
-    | 'plane'
-    | 'play'
-    | 'postcard_heart'
-    | 'pram'
-    | 'qr_code'
-    | 'receipt'
-    | 'rectangle_on_dashed_rectangle'
-    | 'rectangle_speech_bubble'
-    | 'rectangle_speech_bubble_on_speech_bubble'
-    | 'rectangle_watches'
-    | 'round_speech_bubble_question'
-    | 'round_watches'
-    | 'ruble_dollar'
-    | 'ruble_plus'
-    | 'safe'
-    | 'safe_diagram'
-    | 'sber'
-    | 'security_camera'
-    | 'shield'
-    | 'shield_check'
-    | 'shield_crossed'
-    | 'shield_lines'
-    | 'shield_lines_circles'
-    | 'spasibo'
-    | 'speaker_waves'
-    | 'square_on_square'
-    | 'star_light'
-    | 'student_hat'
-    | 'telephone'
-    | 'thermometer'
-    | 'trash'
-    | 'tv_set'
-    | 'umbrella'
-    | 'user'
-    | 'user_cross'
-    | 'user_on_user'
-    | 'vibration'
-    | 'wallet'
-    | 'watering_can'
-    | 'waves'
-    | 'waves_cross'
-    | 'whistle'
-    | 'wrench'
-    | 'yule';
-/**
- * Режим растягивания изображения
- */
-export type ScaleMode =
-    | 'scale_aspect_fill'
-    | 'scale_aspect_fit'
-    | 'center'
-    | 'top'
-    | 'bottom'
-    | 'left'
-    | 'right'
-    | 'top_left'
-    | 'top_right'
-    | 'bottom_left'
-    | 'bottom_right';
-/**
- * Цвет фона см https://www.figma.com/file/yaeE0lLDWMBKKLXuDHgq1p/SmartX-Styles?node-id=19%3A15
- */
-export type Color =
-    | 'solid_black'
-    | 'solid_white'
-    | 'solid_transparent'
-    | 'solid_disabled'
-    | 'solid_brand'
-    | 'solid_warning'
-    | 'solid_critical'
-    | 'solid_action'
-    | 'liquid_60'
-    | 'liquid_50'
-    | 'liquid_40'
-    | 'liquid_30'
-    | 'liquid_20'
-    | 'liquid_10';
-export type Action = DeepLinkAction | SendContactPhone | ServerAction | TextAction;
-export type Actions = Action[];
-/**
- * Ширина контента в терминах сеток. https://www.figma.com/file/L6AlpyUxFuumKEEbh4ADIh/🤖SD-Styles?node-id=26%3A17 https://www.figma.com/file/c6KZ9jIXTB5zEmyTfA2Akv/Message-Based?node-id=1446%3A0
- */
-export type ContentWidth = 'xsmall' | 'small' | 'medium' | 'large' | 'resizable';
-/**
- * Горизонтальная галерея
- */
-export type GalleryCard = CardBase & {
-    items: Array<GalleryItem>;
-    bottom_text?: TextView2;
-    type: 'gallery_card';
-    [k: string]: unknown;
-};
-export type GalleryItem = MediaGalleryItem | GallerySearchContentItem | GalleryMoreButtonItem;
-/**
- * Стиль текста, см https://www.figma.com/file/D0AfmXWT5KBPiWFNg2IZoF/SmartX-Typeface?node-id=0%3A1
- */
-export type Typeface =
-    | 'headline1'
-    | 'headline2'
-    | 'headline3'
-    | 'title1'
-    | 'title2'
-    | 'body1'
-    | 'body2'
-    | 'body3'
-    | 'text1'
-    | 'paragraphText1'
-    | 'paragraphText2'
-    | 'footnote1'
-    | 'footnote2'
-    | 'button1'
-    | 'button2'
-    | 'caption';
-/**
- * Цвет текста см https://www.figma.com/file/yaeE0lLDWMBKKLXuDHgq1p/SmartX-Styles?node-id=239%3A64
- */
-export type TypeColor = 'default' | 'secondary' | 'tertiary' | 'inverse' | 'brand' | 'warning' | 'critical' | 'link';
-/**
- * Адрес графического ресурса, может быть задан в пространстве интернет или в локальном пространстве клиента, url имеет больший приоритет чем local
- */
-export type ImageAddress =
-    | {
-          /**
-           * Тип ресурса.
-           */
-          type: 'url';
-          /**
-           * Адрес картинки в интернете
-           */
-          url: string;
-          /**
-           * Hash катинки в интернете(опционален)
-           */
-          hash?: string;
-          placeholder?: LocalImageIdentificator1;
-          [k: string]: unknown;
-      }
-    | {
-          /**
-           * Тип ресурса.
-           */
-          type: 'local';
-          identificator: LocalImageIdentificator2;
-          [k: string]: unknown;
-      };
-/**
- * Заглушка, которую показываем, когда нет изображения или оно невалидно(опционален)
- */
-export type LocalImageIdentificator1 =
-    | 'add_contactless_payment'
-    | 'apple_wallet'
-    | 'arrow_right_chevron_rectangle'
-    | 'arrow_right_trace_line'
-    | 'arrows_forward_back'
-    | 'atm_banknote'
-    | 'bag'
-    | 'ballot_paper'
-    | 'bar_graph'
-    | 'basket'
-    | 'beach_umbrella'
-    | 'bell'
-    | 'bell_crossed'
-    | 'between_accounts'
-    | 'boarding_pass'
-    | 'bolt'
-    | 'box'
-    | 'building'
-    | 'calculator'
-    | 'camera'
-    | 'candy'
-    | 'capsule'
-    | 'car'
-    | 'card'
-    | 'card_amex'
-    | 'card_amex_black'
-    | 'card_amex_gold'
-    | 'card_blocked'
-    | 'card_chevron_left'
-    | 'card_ghost'
-    | 'card_maestro_null'
-    | 'card_mc'
-    | 'card_mc_black'
-    | 'card_mc_gold'
-    | 'card_mc_null'
-    | 'card_mir'
-    | 'card_mir_black'
-    | 'card_mir_gold'
-    | 'card_mir_null'
-    | 'card_momentum'
-    | 'card_on_card'
-    | 'card_plus'
-    | 'card_pro100_null'
-    | 'card_uek'
-    | 'card_viewfinder'
-    | 'card_visa_gold'
-    | 'card_visa_black'
-    | 'card_visa'
-    | 'card_visa_infinite_sber'
-    | 'card_visa_infinite'
-    | 'card_visa_digital'
-    | 'card_visa_null'
-    | 'cards_carousel'
-    | 'case'
-    | 'case_diagram'
-    | 'chess_piece'
-    | 'circle_cross'
-    | 'circle_plus'
-    | 'clock'
-    | 'coat_of_arms'
-    | 'coin'
-    | 'coin_heart'
-    | 'color_spasibo'
-    | 'color_total'
-    | 'counterclock_wise'
-    | 'crown'
-    | 'digital_banknotes'
-    | 'document'
-    | 'document_checkmark'
-    | 'document_magnifying_glass'
-    | 'document_on_document'
-    | 'drop'
-    | 'ellipsis'
-    | 'eye'
-    | 'eye_crossed'
-    | 'face'
-    | 'face_id'
-    | 'falling_coin'
-    | 'file_pdf'
-    | 'fingerprint'
-    | 'flag'
-    | 'flame'
-    | 'flasher'
-    | 'funnel'
-    | 'gamepad'
-    | 'gear'
-    | 'gift'
-    | 'globe'
-    | 'hammer'
-    | 'handset'
-    | 'heart'
-    | 'heart_hand'
-    | 'house'
-    | 'house_handset'
-    | 'house_percent'
-    | 'info'
-    | 'key'
-    | 'list'
-    | 'lock_closed'
-    | 'lock_opened'
-    | 'magnit'
-    | 'mail'
-    | 'mail_eye'
-    | 'mail_waves'
-    | 'man_badge'
-    | 'man_wheelchair'
-    | 'map'
-    | 'mask'
-    | 'mindmap'
-    | 'mobile'
-    | 'mobile_plus'
-    | 'mobile_plus_alt'
-    | 'mobile_waves'
-    | 'necklace'
-    | 'note'
-    | 'notebook_waves'
-    | 'open_book'
-    | 'other'
-    | 'paper_tray'
-    | 'parking'
-    | 'passport'
-    | 'paw'
-    | 'pencil_line'
-    | 'percent'
-    | 'pfm'
-    | 'photo'
-    | 'pie_chart'
-    | 'placeholder'
-    | 'plane'
-    | 'play'
-    | 'postcard_heart'
-    | 'pram'
-    | 'qr_code'
-    | 'receipt'
-    | 'rectangle_on_dashed_rectangle'
-    | 'rectangle_speech_bubble'
-    | 'rectangle_speech_bubble_on_speech_bubble'
-    | 'rectangle_watches'
-    | 'round_speech_bubble_question'
-    | 'round_watches'
-    | 'ruble_dollar'
-    | 'ruble_plus'
-    | 'safe'
-    | 'safe_diagram'
-    | 'sber'
-    | 'security_camera'
-    | 'shield'
-    | 'shield_check'
-    | 'shield_crossed'
-    | 'shield_lines'
-    | 'shield_lines_circles'
-    | 'spasibo'
-    | 'speaker_waves'
-    | 'square_on_square'
-    | 'star_light'
-    | 'student_hat'
-    | 'telephone'
-    | 'thermometer'
-    | 'trash'
-    | 'tv_set'
-    | 'umbrella'
-    | 'user'
-    | 'user_cross'
-    | 'user_on_user'
-    | 'vibration'
-    | 'wallet'
-    | 'watering_can'
-    | 'waves'
-    | 'waves_cross'
-    | 'whistle'
-    | 'wrench'
-    | 'yule';
-/**
- * Адрес картинки в локальном пространстве клиента https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=19%3A9
- */
-export type LocalImageIdentificator2 =
-    | 'add_contactless_payment'
-    | 'apple_wallet'
-    | 'arrow_right_chevron_rectangle'
-    | 'arrow_right_trace_line'
-    | 'arrows_forward_back'
-    | 'atm_banknote'
-    | 'bag'
-    | 'ballot_paper'
-    | 'bar_graph'
-    | 'basket'
-    | 'beach_umbrella'
-    | 'bell'
-    | 'bell_crossed'
-    | 'between_accounts'
-    | 'boarding_pass'
-    | 'bolt'
-    | 'box'
-    | 'building'
-    | 'calculator'
-    | 'camera'
-    | 'candy'
-    | 'capsule'
-    | 'car'
-    | 'card'
-    | 'card_amex'
-    | 'card_amex_black'
-    | 'card_amex_gold'
-    | 'card_blocked'
-    | 'card_chevron_left'
-    | 'card_ghost'
-    | 'card_maestro_null'
-    | 'card_mc'
-    | 'card_mc_black'
-    | 'card_mc_gold'
-    | 'card_mc_null'
-    | 'card_mir'
-    | 'card_mir_black'
-    | 'card_mir_gold'
-    | 'card_mir_null'
-    | 'card_momentum'
-    | 'card_on_card'
-    | 'card_plus'
-    | 'card_pro100_null'
-    | 'card_uek'
-    | 'card_viewfinder'
-    | 'card_visa_gold'
-    | 'card_visa_black'
-    | 'card_visa'
-    | 'card_visa_infinite_sber'
-    | 'card_visa_infinite'
-    | 'card_visa_digital'
-    | 'card_visa_null'
-    | 'cards_carousel'
-    | 'case'
-    | 'case_diagram'
-    | 'chess_piece'
-    | 'circle_cross'
-    | 'circle_plus'
-    | 'clock'
-    | 'coat_of_arms'
-    | 'coin'
-    | 'coin_heart'
-    | 'color_spasibo'
-    | 'color_total'
-    | 'counterclock_wise'
-    | 'crown'
-    | 'digital_banknotes'
-    | 'document'
-    | 'document_checkmark'
-    | 'document_magnifying_glass'
-    | 'document_on_document'
-    | 'drop'
-    | 'ellipsis'
-    | 'eye'
-    | 'eye_crossed'
-    | 'face'
-    | 'face_id'
-    | 'falling_coin'
-    | 'file_pdf'
-    | 'fingerprint'
-    | 'flag'
-    | 'flame'
-    | 'flasher'
-    | 'funnel'
-    | 'gamepad'
-    | 'gear'
-    | 'gift'
-    | 'globe'
-    | 'hammer'
-    | 'handset'
-    | 'heart'
-    | 'heart_hand'
-    | 'house'
-    | 'house_handset'
-    | 'house_percent'
-    | 'info'
-    | 'key'
-    | 'list'
-    | 'lock_closed'
-    | 'lock_opened'
-    | 'magnit'
-    | 'mail'
-    | 'mail_eye'
-    | 'mail_waves'
-    | 'man_badge'
-    | 'man_wheelchair'
-    | 'map'
-    | 'mask'
-    | 'mindmap'
-    | 'mobile'
-    | 'mobile_plus'
-    | 'mobile_plus_alt'
-    | 'mobile_waves'
-    | 'necklace'
-    | 'note'
-    | 'notebook_waves'
-    | 'open_book'
-    | 'other'
-    | 'paper_tray'
-    | 'parking'
-    | 'passport'
-    | 'paw'
-    | 'pencil_line'
-    | 'percent'
-    | 'pfm'
-    | 'photo'
-    | 'pie_chart'
-    | 'placeholder'
-    | 'plane'
-    | 'play'
-    | 'postcard_heart'
-    | 'pram'
-    | 'qr_code'
-    | 'receipt'
-    | 'rectangle_on_dashed_rectangle'
-    | 'rectangle_speech_bubble'
-    | 'rectangle_speech_bubble_on_speech_bubble'
-    | 'rectangle_watches'
-    | 'round_speech_bubble_question'
-    | 'round_watches'
-    | 'ruble_dollar'
-    | 'ruble_plus'
-    | 'safe'
-    | 'safe_diagram'
-    | 'sber'
-    | 'security_camera'
-    | 'shield'
-    | 'shield_check'
-    | 'shield_crossed'
-    | 'shield_lines'
-    | 'shield_lines_circles'
-    | 'spasibo'
-    | 'speaker_waves'
-    | 'square_on_square'
-    | 'star_light'
-    | 'student_hat'
-    | 'telephone'
-    | 'thermometer'
-    | 'trash'
-    | 'tv_set'
-    | 'umbrella'
-    | 'user'
-    | 'user_cross'
-    | 'user_on_user'
-    | 'vibration'
-    | 'wallet'
-    | 'watering_can'
-    | 'waves'
-    | 'waves_cross'
-    | 'whistle'
-    | 'wrench'
-    | 'yule';
-/**
- * Ширина контента в терминах сеток. https://www.figma.com/file/L6AlpyUxFuumKEEbh4ADIh/🤖SD-Styles?node-id=26%3A17 https://www.figma.com/file/c6KZ9jIXTB5zEmyTfA2Akv/Message-Based?node-id=1446%3A0
- */
-export type ContentWidth1 = 'xsmall' | 'small' | 'medium' | 'large' | 'resizable';
-/**
- * Цвет фона см https://www.figma.com/file/yaeE0lLDWMBKKLXuDHgq1p/SmartX-Styles?node-id=19%3A15
- */
-export type Color1 =
-    | 'solid_black'
-    | 'solid_white'
-    | 'solid_transparent'
-    | 'solid_disabled'
-    | 'solid_brand'
-    | 'solid_warning'
-    | 'solid_critical'
-    | 'solid_action'
-    | 'liquid_60'
-    | 'liquid_50'
-    | 'liquid_40'
-    | 'liquid_30'
-    | 'liquid_20'
-    | 'liquid_10';
-/**
- * Двумерная карточка (сетка)
- */
-export type GridCard = CardBase & {
-    /**
-     * Количество ячеек
-     */
-    items: GridItem[];
-    type: 'grid_card';
-    /**
-     * Количество столбцов
-     */
-    columns?: number;
-    item_width?: ContentWidth2;
-    /**
-     * Высота карточки
-     */
-    item_height?: 'fixed' | 'resizable';
-    [k: string]: unknown;
-};
-export type GridItem = GreetingGridItem;
-/**
- * Ширина контента в терминах сеток. https://www.figma.com/file/L6AlpyUxFuumKEEbh4ADIh/🤖SD-Styles?node-id=26%3A17 https://www.figma.com/file/c6KZ9jIXTB5zEmyTfA2Akv/Message-Based?node-id=1446%3A0
- */
-export type ContentWidth2 = 'xsmall' | 'small' | 'medium' | 'large' | 'resizable';
-/**
- * Карточка с вертикальным списком ячеек.
- */
-export type ListCard = CardBase & {
-    cells: CellView[];
-    type: 'list_card';
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-};
-export type CellView =
-    | LeftRightCellView
-    | TextCellView
-    | ButtonCellView
-    | ImageCellView
-    | DetailsCellView
-    | WeatherCellView;
-export type LeftView = SimpleLeftView | FastAnswerLeftView;
-/**
- * Гравити ui элемента по вертикали.
- */
-export type VerticalGravity = 'top' | 'center' | 'bottom';
-export type RightView = DisclosureRightView | DetailRightView;
-/**
- * Стиль текста, см https://www.figma.com/file/D0AfmXWT5KBPiWFNg2IZoF/SmartX-Typeface?node-id=0%3A1
- */
-export type Typeface1 =
-    | 'headline1'
-    | 'headline2'
-    | 'headline3'
-    | 'title1'
-    | 'title2'
-    | 'body1'
-    | 'body2'
-    | 'body3'
-    | 'text1'
-    | 'paragraphText1'
-    | 'paragraphText2'
-    | 'footnote1'
-    | 'footnote2'
-    | 'button1'
-    | 'button2'
-    | 'caption';
-export type Actions1 = [Action, ...Action[]];
-
-/**
- * Свойства любой карточки.
- */
-export interface CardBase {
-    paddings?: Paddings;
-    /**
-     * Свойство, говорящее о том, что карточка может стать визуально задизейблена.
-     */
-    can_be_disabled?: boolean;
-    [k: string]: unknown;
-}
-/**
- * Объект для описания отступов ui компонента.
- */
-export interface Paddings {
-    /**
-     * Возможные размеры отступов
-     */
-    left?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    top?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    right?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    bottom?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    [k: string]: unknown;
-}
-/**
- * Изображение из интернета. Ширина изображения будет соответствовать всей ширине карточке.
- */
-export interface ImageView {
-    /**
-     * Адрес изображения в интернет
-     */
-    url: string;
-    /**
-     * Hash катинки в интернете(опционален)
-     */
-    hash?: string;
-    placeholder?: LocalImageIdentificator;
-    scale_mode?: ScaleMode;
-    /**
-     * Высота контейнера под картинку, указывается в платформонезависимых поинтах
-     */
-    height?: number;
-    placeholder_color?: Color;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    size?: ContentSize;
-    [k: string]: unknown;
-}
-/**
- * Действие, которое обозначает отправку номера телефона указанного контакта
- */
-export interface SendContactPhone {
-    /**
-     * Тип действия.
-     */
-    type: 'send_contact_phone';
-    /**
-     * Id контакта
-     */
-    send_contact_phone: number;
-    /**
-     * Может содержать подстроку [[placeholder]]. Клиент должен подставить в строку шаблона вместо [[placeholder]] номер телефона контакта имеющий id указаный в send_contact_phone
-     */
-    template?: string;
-    [k: string]: unknown;
-}
-/**
- * Размера изображения. Если есть size, то height игнорируется
- */
-export interface ContentSize {
-    width: ContentWidth;
-    /**
-     * Отношение высоты контента к ширине
-     */
-    aspect_ratio: number;
-    [k: string]: unknown;
-}
-/**
- * Элемент галереи состоящий из изображения и двух текстовых полей
- */
-export interface MediaGalleryItem {
-    /**
-     * Тип элемента галереи
-     */
-    type: 'media_gallery_item';
-    image: ImageView;
-    top_text?: TextView;
-    bottom_text?: TextView;
-    margins?: Margins1;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Текстовый элемент
- */
-export interface TextView {
-    /**
-     * Текст, который нужно отобразить
-     */
-    text: string;
-    typeface: Typeface;
-    text_color: TypeColor;
-    margins?: Margins;
-    /**
-     * Максимальное количество строк. 0 - не ограничено.
-     */
-    max_lines?: number;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Объект для описания отступов ui компонента.
- */
-export interface Margins {
-    /**
-     * Возможные размеры отступов
-     */
-    left?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    top?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    right?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    bottom?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    [k: string]: unknown;
-}
-/**
- * Объект для описания отступов ui компонента.
- */
-export interface Margins1 {
-    /**
-     * Возможные размеры отступов
-     */
-    left?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    top?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    right?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    bottom?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    [k: string]: unknown;
-}
-/**
- * Элемент карточки галереи с заголовком, ссылкой и описанием
- */
-export interface GallerySearchContentItem {
-    /**
-     * Тип элемента отдельной карточки
-     */
-    type: 'gallery_search_content_item';
-    title_text: TextView;
-    image?: ImageAddress;
-    link_text: TextView;
-    description_text?: TextView;
-    paddings?: Paddings1;
-    actions: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    width?: ContentWidth1;
-    [k: string]: unknown;
-}
-/**
- * Объект для описания отступов ui компонента.
- */
-export interface Paddings1 {
-    /**
-     * Возможные размеры отступов
-     */
-    left?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    top?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    right?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    bottom?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    [k: string]: unknown;
-}
-/**
- * Кнопка "Все результаты" в карточке галереи
- */
-export interface GalleryMoreButtonItem {
-    /**
-     * Тип элемента кнопки всех результатов
-     */
-    type: 'gallery_more_button_item';
-    bottom_text: TextView1;
-    icon: IconView;
-    actions: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Текстовый элемент
- */
-export interface TextView1 {
-    /**
-     * Текст, который нужно отобразить
-     */
-    text: string;
-    typeface: Typeface;
-    text_color: TypeColor;
-    margins?: Margins;
-    /**
-     * Максимальное количество строк. 0 - не ограничено.
-     */
-    max_lines?: number;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Иконка
- */
-export interface IconView {
-    address: ImageAddress;
-    size: IconSize;
-    margins?: Margins;
-    tint_color?: Color1;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Размер изображения
- */
-export interface IconSize {
-    /**
-     * Возможные значения размеров изображения.
-     */
-    width: 'xsmall' | 'small' | 'medium' | 'large' | 'xlarge';
-    /**
-     * Возможные значения размеров изображения.
-     */
-    height: 'xsmall' | 'small' | 'medium' | 'large' | 'xlarge';
-    [k: string]: unknown;
-}
-/**
- * Текстовый элемент
- */
-export interface TextView2 {
-    /**
-     * Текст, который нужно отобразить
-     */
-    text: string;
-    typeface: Typeface;
-    text_color: TypeColor;
-    margins?: Margins;
-    /**
-     * Максимальное количество строк. 0 - не ограничено.
-     */
-    max_lines?: number;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Элемент карточки приветствия, состоящий из двух текстовых полей и опционального изображения
- */
-export interface GreetingGridItem {
-    /**
-     * Тип элемента карточки приветствия
-     */
-    type: 'greeting_grid_item';
-    background_image?: ImageView;
-    top_text: TextView;
-    bottom_text: TextView;
-    paddings?: Paddings2;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Объект для описания отступов ui компонента.
- */
-export interface Paddings2 {
-    /**
-     * Возможные размеры отступов
-     */
-    left?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    top?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    right?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    /**
-     * Возможные размеры отступов
-     */
-    bottom?: '0x' | '1x' | '2x' | '4x' | '5x' | '6x' | '8x' | '9x' | '10x' | '12x' | '16x';
-    [k: string]: unknown;
-}
-export interface LeftRightCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'left_right_cell_view';
-    left: LeftView;
-    right?: RightView;
-    divider?: DividerView;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-/**
- * Горизонтальный упорядоченный блок содержащий icon и texts
- */
-export interface SimpleLeftView {
-    /**
-     * Тип левого элемента ячейки.
-     */
-    type: 'simple_left_view';
-    icon?: IconView1;
-    /**
-     * Вертикальный упорядоченный список содержащий title, subtitle, caption, https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=19%3A5641
-     */
-    texts?: {
-        title?: TextView;
-        subtitle?: TextView;
-        caption?: TextView;
-        margins?: Margins;
-        [k: string]: unknown;
-    };
-    icon_vertical_gravity?: VerticalGravity;
-    title?: TextView3;
-    [k: string]: unknown;
-}
-/**
- * Иконка
- */
-export interface IconView1 {
-    address: ImageAddress;
-    size: IconSize;
-    margins?: Margins;
-    tint_color?: Color1;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Текстовый элемент
- */
-export interface TextView3 {
-    /**
-     * Текст, который нужно отобразить
-     */
-    text: string;
-    typeface: Typeface;
-    text_color: TypeColor;
-    margins?: Margins;
-    /**
-     * Максимальное количество строк. 0 - не ограничено.
-     */
-    max_lines?: number;
-    actions?: Actions;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    [k: string]: unknown;
-}
-/**
- * Вертикальный упорядоченный список, состоит из label, icon_and_value https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=227%3A3670
- */
-export interface FastAnswerLeftView {
-    /**
-     * Тип левого элемента ячейки.
-     */
-    type: 'fast_answer_left_view';
-    label?: TextView;
-    /**
-     * Горизонтальный упорядоченный список содержит icon и value, https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=227%3A3670
-     */
-    icon_and_value?: {
-        icon?: IconView1;
-        value?: TextView;
-        margins?: Margins;
-        required?: ['value'];
-        [k: string]: unknown;
-    };
-    margins?: Margins;
-    [k: string]: unknown;
-}
-/**
- * Угловая скобка. Соответствует UITableViewCell.AccessoryType.disclosureIndicator
- */
-export interface DisclosureRightView {
-    /**
-     * Тип правого элемента ячейки.
-     */
-    type: 'disclosure_right_view';
-    margins?: Margins;
-    [k: string]: unknown;
-}
-/**
- * Вертикальный упорядоченный блок, содержащий опциональные detail и info_box https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=19%3A5686
- */
-export interface DetailRightView {
-    /**
-     * Тип правого элемента ячейки.
-     */
-    type: 'detail_right_view';
-    detail?: TextView;
-    /**
-     * Горизнотальный упорядоченный блок, содержащий опциональные info и icon https://www.figma.com/file/MQHBgPkW4dqXmI2549WioI/Dev?node-id=19%3A5686
-     */
-    info_and_icon?: {
-        info?: TextView;
-        icon?: IconView1;
-        margins?: Margins;
-        [k: string]: unknown;
-    };
-    margins?: Margins;
-    vertical_gravity?: VerticalGravity;
-    /**
-     * Расположение detail относительно info_and_icon
-     */
-    detail_position?: 'top' | 'bottom';
-    [k: string]: unknown;
-}
-/**
- * Divider/Separator. Тонка полоска, разделяющая ячейки
- */
-export interface DividerView {
-    /**
-     * Стиль дивайдера.
-     */
-    style: 'default' | 'read_only';
-    /**
-     * Возможные размеры сепаратора/дивайдера https://www.figma.com/file/sC1HJg3HRcNjaG3rcd4b3V/Templates?node-id=289%3A3210
-     */
-    size: 'd1' | 'd2' | 'd3' | 'd4' | 'd5';
-    [k: string]: unknown;
-}
-export interface TextCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'text_cell_view';
-    content: TextView;
-    divider?: DividerView;
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-export interface ButtonCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'button_cell_view';
-    content: ButtonView;
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-/**
- * Кнопки для универсальных карточек
- */
-export interface ButtonView {
-    /**
-     * Текст кнопки.
-     */
-    text: string;
-    typeface?: Typeface1;
-    /**
-     * Стиль кнопки.
-     */
-    style?: 'default' | 'transparent';
-    /**
-     * Тип кнопки.
-     */
-    type?: 'accept' | 'disabled' | 'negative';
-    actions: Actions1;
-    /**
-     * Id для отправки в метрику
-     */
-    log_id?: string;
-    margins?: Margins;
-    [k: string]: unknown;
-}
-export interface ImageCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'image_cell_view';
-    content: ImageView;
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-export interface DetailsCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'details_cell_view';
-    details: [DetailRightView, ...DetailRightView[]];
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-export interface WeatherCellView {
-    /**
-     * Тип ячейки
-     */
-    type: 'weather_cell_view';
-    content: [WeatherView, ...WeatherView[]];
-    paddings?: Paddings;
-    [k: string]: unknown;
-}
-/**
- * Строка карточки погоды на неделю
- */
-export interface WeatherView {
-    weekday: string;
-    date: string;
-    day: string;
-    night: string;
-    icon: IconView1;
-    [k: string]: unknown;
-}
-/**
- * Виджет с заголовком раздела и действием
- */
-export interface TitleCard {
-    cell: LeftRightCellView;
-    type: 'title_card';
-    [k: string]: unknown;
-}
