+++ packages/types/src/index.ts
@@ -1,8 +1,6 @@
+export * from './i18n';
 export * from './request';
 export * from './response';
 export * from './salute';
-export * from './global';
-export * from './i18n';
 export * from './storage';
-export * from './card';
-export * from './bubble';
+export * from './systemMessage';
